BXC-3284 move to web services app
